+++ engine/renderer/vulkan/raster/renderer_rasterization.h
@@ -13,7 +13,7 @@ namespace lotus
         virtual Task<> Init() override;
 
         virtual Task<> drawFrame() override;
-        virtual void populateAccelerationStructure(TopLevelAccelerationStructure*, BottomLevelAccelerationStructure*, const glm::mat3x4&, uint64_t, uint32_t, uint32_t) override {}
+        virtual void populateAccelerationStructure(TopLevelAccelerationStructure*, BottomLevelAccelerationStructure*, const glm::mat3x4&, uint32_t, uint32_t, uint32_t) override {}
 
         virtual void initEntity(EntityInitializer*, Engine*) override;
         virtual void drawEntity(EntityInitializer*, Engine*) override;
